convert/requirements-convert.txt
# Conversion toolchain adjusted for Python 3.12
# NOTE: TensorFlow 2.15 wheels are not published for 3.12 (at time of writing), so we cannot use onnx-tf + tfjs direct path here.
# Instead we will: PyTorch -> ONNX -> (later run TF conversion in an external py310 env) OR use direct ONNX -> TFJS route (experimental).

numpy==1.26.4
torch==2.3.1
torchvision==0.18.1
timm==1.0.7
onnx==1.15.0
onnxruntime==1.17.3
tensorflowjs==4.20.0

# Optional (only if you later switch to Python 3.10/3.11):
# tensorflow==2.15.1
# tensorflow-addons==0.23.0
# onnx-tf==1.11.0
